Moved methods to scheduledDashboard component
